GUI 수정, 코드 보완

+++ src/GUIBasedJavaIDE.java
@@ -4,11 +4,10 @@ import java.awt.event.*;
 import java.io.*;
 
 public class GUIBasedJavaIDE extends JFrame {
-  private JTextField fileOpenField;
-  private JTextField fileSaveField;
-  private JTextArea editingWindow;
-  private JTextArea resultWindow;
-  private JButton openBtn, saveBtn, compileBtn, saveErrorsBtn, deleteBtn, clearBtn;
+  private final JTextField fileOpenField;
+  private final JTextField fileSaveField;
+  private final JTextArea editingWindow;
+  private final JTextArea resultWindow;
   private String fName = null;
   private StringBuilder errorContent = null;
 
@@ -16,116 +15,126 @@ public class GUIBasedJavaIDE extends JFrame {
     // GUI 기본 설정
     setTitle("My Java IDE GUI");
     setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-    setSize(600, 500);
-    setLayout(new BorderLayout());
+    setSize(800, 720);
+    Container c = getContentPane();
+    c.setLayout(new BorderLayout());
 
     // 그리드백 레이아웃을 탑재한 컨텐트펜 생성
-    JPanel contentPane = new JPanel();
-    contentPane.setLayout(new GridBagLayout());
+    JPanel contentPane = new JPanel(new GridBagLayout());
+    contentPane.setBorder(BorderFactory.createEmptyBorder(0, 120, 0, 120));
     // 그리드백 레이아웃 내 컴포넌트 배치 설정을 관리하는 객체 생성, 여백 설정
-    GridBagConstraints gbc = new GridBagConstraints();
-    gbc.insets = new Insets(5, 5, 5, 5);
+    GridBagConstraints contentGBC = new GridBagConstraints();
+    contentGBC.insets = new Insets(5, 5, 5, 5);
 
     // 파일 불러오기를 위한 텍스트필드의 위치 및 너비 설정
     fileOpenField = new JTextField(40);
-    gbc.gridx = 0;
-    gbc.gridy = 0;
-    gbc.gridwidth = 5;
-    contentPane.add(fileOpenField, gbc);
+    contentGBC.gridx = 0;
+    contentGBC.gridy = 0;
+    contentGBC.gridwidth = 5;
+    contentGBC.weightx = 1;
+    contentGBC.fill = GridBagConstraints.HORIZONTAL;
+    contentPane.add(fileOpenField, contentGBC);
 
     // 불러오기 버튼의 위치 및 너비, 이벤트 처리 리스너 설정
-    openBtn = new JButton("Open");
+    JButton openBtn = new JButton("Open");
     openBtn.addActionListener(new OpenListener());
-    gbc.gridx = 5;
-    gbc.gridwidth = 1;
-    contentPane.add(openBtn, gbc);
+    contentGBC.gridx = 5;
+    contentGBC.gridwidth = 1;
+    contentGBC.weightx = 0;
+    contentGBC.fill = GridBagConstraints.NONE;
+    contentPane.add(openBtn, contentGBC);
 
     // 파일 저장하기를 위한 텍스트필드의 위치 및 너비 설정
     fileSaveField = new JTextField(40);
-    gbc.gridx = 0;
-    gbc.gridy = 1;
-    gbc.gridwidth = 5;
-    contentPane.add(fileSaveField, gbc);
+    contentGBC.gridx = 0;
+    contentGBC.gridy = 1;
+    contentGBC.gridwidth = 5;
+    contentGBC.weightx = 1;
+    contentGBC.fill = GridBagConstraints.HORIZONTAL;
+    contentPane.add(fileSaveField, contentGBC);
 
     // 저장하기 버튼의 위치 및 너비, 이벤트 처리 리스너 설정
-    saveBtn = new JButton("Save");
+    JButton saveBtn = new JButton("Save");
     saveBtn.addActionListener(new SaveListener());
-    gbc.gridx = 5;
-    gbc.gridwidth = 1;
-    contentPane.add(saveBtn, gbc);
+    contentGBC.gridx = 5;
+    contentGBC.gridwidth = 1;
+    contentGBC.weightx = 0;
+    contentGBC.fill = GridBagConstraints.NONE;
+    contentPane.add(saveBtn, contentGBC);
 
     // 코드 편집을 위한 Editing window 텍스트 에리어의 스크롤, 위치, 너비, 높이비율, 공간 채우기 설정
     editingWindow = new JTextArea();
     JScrollPane editingScrollPane = new JScrollPane(editingWindow);
-    gbc.gridx = 0;
-    gbc.gridy = 3;
-    gbc.gridwidth = 6;
-    gbc.weighty = 8;
-    gbc.fill = GridBagConstraints.BOTH;
-    contentPane.add(editingScrollPane, gbc);
+    contentGBC.gridx = 0;
+    contentGBC.gridy = 3;
+    contentGBC.gridwidth = 6;
+    contentGBC.weighty = 8;
+    contentGBC.fill = GridBagConstraints.BOTH;
+    contentPane.add(editingScrollPane, contentGBC);
 
     // 4개 버튼 배치를 위한 패널 생성
-    JPanel buttonPannel = new JPanel(new GridLayout(1, 4, 5, 5));
+    JPanel buttonPanel = new JPanel(new GridLayout(1, 4, 5, 5));
 
     // 컴파일 버튼, 에러 파일 저장 버튼, 자바 파일 삭제 버튼, 클리어 버튼 생성 및 이벤트 처리 리스너 설정
-    compileBtn = new JButton("Compile");
-    compileBtn.addActionListener(new compileListener());
-    saveErrorsBtn = new JButton("Save Errors");
-    saveErrorsBtn.addActionListener(new saveErrorsListener());
-    deleteBtn = new JButton("Delete");
-    deleteBtn.addActionListener(new deleteListener());
-    clearBtn = new JButton("Clear");
-    clearBtn.addActionListener(new clearListener());
+    JButton compileBtn = new JButton("Compile");
+    compileBtn.addActionListener(new CompileListener());
+    JButton saveErrorsBtn = new JButton("Save Errors");
+    saveErrorsBtn.addActionListener(new SaveErrorsListener());
+    JButton deleteBtn = new JButton("Delete");
+    deleteBtn.addActionListener(new DeleteListener());
+    JButton clearBtn = new JButton("Clear");
+    clearBtn.addActionListener(new ClearListener());
 
     // 버튼 패널에 버튼 탑재
-    buttonPannel.add(compileBtn);
-    buttonPannel.add(saveErrorsBtn);
-    buttonPannel.add(deleteBtn);
-    buttonPannel.add(clearBtn);
+    buttonPanel.add(compileBtn);
+    buttonPanel.add(saveErrorsBtn);
+    buttonPanel.add(deleteBtn);
+    buttonPanel.add(clearBtn);
 
     // 패널의 위치 및 높이비율, 공간 채우기 설정
-    gbc.gridy = 4;
-    gbc.weighty = 1;
-    gbc.fill = GridBagConstraints.HORIZONTAL;
-    contentPane.add(buttonPannel, gbc);
+    contentGBC.gridy = 4;
+    contentGBC.weighty = 1;
+    contentGBC.fill = GridBagConstraints.HORIZONTAL;
+    contentPane.add(buttonPanel, contentGBC);
 
     // 동작 수행 결과를 나태니기 위한 Result window 텍스트 에리어의 편집 가능 여부, 스크롤, 위치, 높이비율, 공간 채우기 설정
     resultWindow = new JTextArea();
     resultWindow.setEditable(false);
     JScrollPane resultScrollPane = new JScrollPane(resultWindow);
-    gbc.gridy = 5;
-    gbc.weighty = 5;
-    gbc.fill = GridBagConstraints.BOTH;
-    contentPane.add(resultScrollPane, gbc);
+    contentGBC.gridy = 5;
+    contentGBC.weighty = 5;
+    contentGBC.fill = GridBagConstraints.BOTH;
+    contentPane.add(resultScrollPane, contentGBC);
 
     // 프레임에 컨텐트펜 탑재
-    add(contentPane, BorderLayout.CENTER);
+    c.add(contentPane, BorderLayout.CENTER);
     setVisible(true);
   }
 
   // 파일 불러오기 버튼의 이벤트 처리 리스너
   private class OpenListener implements ActionListener {
     public void actionPerformed(ActionEvent event) {
-      // 파일 이름 입력 받아서 저장, 에러내용 초기화
-      fName = fileOpenField.getText();
-      errorContent = null;
-      File file = new File("javafile/" + fName);
+      // 파일 이름 읽어오기
+      String openName = fileOpenField.getText();
+      File file = new File("javafile/" + openName);
       // 파일 존재 여부 및 확장자 검사
-      if (file.exists() && fName.endsWith(".java")) {
+      if (file.exists() && openName.endsWith(".java")) {
         try {
           // 자바 파일을 읽어들일 입력 버퍼 생성
           BufferedReader reader = new BufferedReader(new FileReader(file));
 
-          // 자바 파일 내용을 Editing Window에 출력
+          // 자바 파일 내용을 Editing Window 에 출력
           String stream;
           editingWindow.setText("");
           while ((stream = reader.readLine()) != null) editingWindow.append(stream + "\n");
 
-          // 버퍼 닫기, 프로그램 타이틀 변경, 파일 입력 필드 초기화, 실행 결과 표시
+          // 버퍼 닫기, 파일 입력 필드 초기화, 실행 결과 표시, 파일 이름 저장, 에러 내용 초기화, 프로그램 타이틀 변경
           reader.close();
-          setTitle("\"" + fName + "\" - My Java IDE GUI");
           fileOpenField.setText("");
-          resultWindow.setText("Java file loaded successfully");
+          resultWindow.setText("Java file loaded successfully.");
+          fName = openName;
+          errorContent = null;
+          setTitle("\"" + openName + "\" - My Java IDE GUI");
         } catch (IOException error) {
           resultWindow.setText("Error: Failed to read the java file!");
         }
@@ -139,40 +148,58 @@ public class GUIBasedJavaIDE extends JFrame {
   // 파일 저장 버튼의 이벤트 처리 리스너
   private class SaveListener implements ActionListener {
     public void actionPerformed(ActionEvent event) {
-// 파일 저장 이름 가져오기
-      String saveFileName = fileSaveField.getText().trim(); // 저장할 파일 이름 가져오기 (공백 제거)
+      // 덮어쓰기 플래그
+      boolean overwrite = false;
+      // 파일 저장 이름 가져오기
+      String saveFileName = fileSaveField.getText().strip(); // 저장할 파일 이름 가져오기 (앞뒤 공백 제거)
 
       // 저장할 파일 이름이 제공되지 않은 경우, 현재 열려 있는 파일에 덮어쓰기
       if (saveFileName.isEmpty()) {
         if (fName == null) { // 현재 열려 있는 파일이 없으면 에러 메시지 출력
-          resultWindow.setText("Error: No file to overwrite. Please provide a file name.");
+          resultWindow.setText("Error: No file to overwrite! Please open a file to overwrite or enter a filename.");
           return;
         }
         saveFileName = fName; // 저장할 파일 이름을 현재 열려 있는 파일 이름으로 설정
+        overwrite = true; // 덮어쓰기 플래그 활성화
+      }
+      // 저장할 파일 이름이 제공되었으나 확장자가 .java 가 아니라면 에러 메시지 출력
+      else if (!saveFileName.endsWith(".java")) {
+        resultWindow.setText("Error: Invalid filename extension! Please enter a filename which ends with '.java'.");
+        return;
       }
 
       // 저장할 파일 생성
       File file = new File("javafile/" + saveFileName);
 
-      try {
-        // 파일에 내용을 저장하기 위한 출력 버퍼 생성
-        BufferedWriter writer = new BufferedWriter(new FileWriter(file));
-        writer.write(editingWindow.getText()); // Editing Window의 텍스트를 파일에 저장
-        writer.flush(); // 버퍼 비우기
-        writer.close(); // 버퍼 닫기
+      // 새로운 파일을 생성하거나 현재 열려있는 파일의 변경 사항을 저장(덮어쓰기)하는 경우
+      if (!file.exists() || overwrite) {
+        try {
+          // 파일에 내용을 저장하기 위한 출력 버퍼 생성
+          BufferedWriter writer = new BufferedWriter(new FileWriter(file));
+          writer.write(editingWindow.getText()); // Editing Window 의 텍스트를 파일에 저장
+          writer.flush(); // 버퍼 비우기
+          writer.close(); // 버퍼 닫기
 
-        // 성공 메시지 출력 및 필드 초기화
-        resultWindow.setText("File saved successfully: " + saveFileName);
-        fileSaveField.setText(""); // 파일 저장 필드 초기화
-        setTitle("\"" + saveFileName + "\" - My Java IDE GUI"); // 타이틀 업데이트
-      } catch (IOException e) {
-        resultWindow.setText("Error: Failed to save the file!");
+          // 성공 메시지 출력 및 필드 초기화
+          resultWindow.setText("File saved successfully: " + saveFileName);
+          fileSaveField.setText(""); // 파일 저장 필드 초기화
+          fName = saveFileName; // 열려있는 파일 교체
+          errorContent = null; // 에러 내용 초기화
+          setTitle("\"" + saveFileName + "\" - My Java IDE GUI"); // 타이틀 업데이트
+        } catch (IOException error) {
+          resultWindow.setText("Error: Failed to save the file! Check if filename contains invalid characters.");
+        }
       }
+      // 동일한 이름의 파일이 이미 존재하는데 파일을 생성하려 경우 에러 메시지 출력
+      else
+        resultWindow.setText("Error: File already exists!\n" +
+                "    1) If you want to create a new file, please enter a not existing filename.\n" +
+                "    2) If you want to overwrite a file, please open the file first and empty name field before saving.");
     }
   }
 
   // 불러온 파일 컴파일 버튼의 이벤트 처리 리스너
-  private class compileListener implements ActionListener {
+  private class CompileListener implements ActionListener {
     public void actionPerformed(ActionEvent event) {
       // 파일 불러오기 여부 검사
       if (fName == null)
@@ -189,15 +216,16 @@ public class GUIBasedJavaIDE extends JFrame {
           boolean hasError = false;
           String stream;
           errorContent = new StringBuilder();
+          resultWindow.setText("");
           while ((stream = stdError.readLine()) != null) {
-            hasError = true; // 오류가 발생했으므로 플래그를 true로 전환
+            hasError = true; // 오류가 발생했으므로 플래그를 true 로 전환
             errorContent.append(stream).append("\n"); // 내용 한줄씩 저장
             resultWindow.append(stream + "\n"); // 한줄씩 출력
           }
 
           // 컴파일 정상 완료 시 출력
           if (!hasError)
-            resultWindow.setText("Compiled successfully...");
+            resultWindow.setText("Compiled successfully: " + fName.replace(".java", ".class"));
 
           // 버퍼 닫기
           stdError.close();
@@ -208,7 +236,7 @@ public class GUIBasedJavaIDE extends JFrame {
     }
   }
 
-  private class saveErrorsListener implements ActionListener {
+  private class SaveErrorsListener implements ActionListener {
     public void actionPerformed(ActionEvent event) {
       // 저장할 오류 내용의 존재 여부 확인
       if (errorContent == null || errorContent.length() == 0)
@@ -233,7 +261,7 @@ public class GUIBasedJavaIDE extends JFrame {
   }
 
   // 삭제 버튼의 이벤트 처리 리스너
-  private class deleteListener implements ActionListener {
+  private class DeleteListener implements ActionListener {
     public void actionPerformed(ActionEvent e) {
       // 파일 불러오기 여부 검사
       if (fName == null)
@@ -243,7 +271,7 @@ public class GUIBasedJavaIDE extends JFrame {
         editingWindow.setText("");
         resultWindow.setText("");
         errorContent = null;
-        // javafile 폴더에서 이름으로 fileName을 갖는 클래스 파일에 대한 객체를 생성
+        // javafile 폴더에서 이름으로 fileName 을 갖는 클래스 파일에 대한 객체를 생성
         File classFile = new File("javafile/" + fName.replace(".java", ".class"));
         // 클래스 파일 존재 시 삭제
         if (classFile.exists()) {
@@ -255,17 +283,18 @@ public class GUIBasedJavaIDE extends JFrame {
         File javaFile = new File("javafile/" + fName);
         // 자바 파일 삭제
         javaFile.delete();
-        resultWindow.append("Java file deleted\n\n");
+        resultWindow.append("Java file deleted.\n\n");
 
-        // 실행 결과 출력 및 프로그램 타이틀 변경
+        // 실행 결과 출력 및 프로그램 타이틀 변경, 파일 이름 초기화
         resultWindow.append("Deletion completed.");
+        fName = null;
         setTitle("My Java IDE GUI");
       }
     }
   }
 
   // 클리어 버튼의 이벤트 처리 리스너
-  private class clearListener implements ActionListener {
+  private class ClearListener implements ActionListener {
     public void actionPerformed(ActionEvent e) {
       // 필드, 윈도우, 파일 이름, 에러 내용 전부 초기화, 프로그램 타이틀 변경
       fileOpenField.setText("");
